ENH: Reorganization of reconstruction module keeping same API calls
